Implementa testes da service e controller e adiciona a anotação PrePersist
Adiciona a anotação PrePersist para confirmar pedido e pegar horário do pedido confirmado ao invés do horário
que o pedido foi feito pelo cliente.

Foi testado teste de integração com o mockMVC e a service com o JUnit

+++ src/main/java/com/pizzaria/pizzaria/pizza/repository/PizzaRepository.java
@@ -3,5 +3,5 @@ package com.pizzaria.pizzaria.pizza.repository;
 import com.pizzaria.pizzaria.pizza.model.Pizza;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PizzaRepository extends JpaRepository<Pizza, Long> {
+public interface PizzaRepository extends JpaRepository<Pizza, Integer> {
 }
